Updated Texture Node

+++ TerraForge3D/src/Generators/CPUNodeEditor/Nodes/TextureNode.h
@@ -24,7 +24,7 @@ public:
 	bool autoTiled;
     bool inv;
     bool npScale;
-	int numTiles;
+	float numTiles;
 	float posi[2];
 	
 private:
